requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.30.0
accelerate>=0.20.0
xformers>=0.0.20
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0
pillow>=10.0.0